+++ app/mailers/notification_mailer.rb
@@ -34,15 +34,12 @@ class NotificationMailer < ActionMailer::Base
       end
 
       if ticket.assignee.nil?
-        message_id = nil
-
         ticket.notified_users.each do |user|
           message = NotificationMailer.new_ticket(ticket, user)
-          message.message_id = message_id
+          message.message_id = ticket.message_id
           message.deliver_now unless EmailAddress.pluck(:email).include?(user.email)
 
-          ticket.message_id = message.message_id
-          message_id = message.message_id
+          ticket.message_id = message.message_id if ticket.message_id.nil?
         end
 
         ticket.save